dom/tests/mochitest/general/test_malformed_url_in_href_throws_type_error.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1788513
-->
<head>
  <title>Test for Bug 1788513</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1788513">Mozilla Bug 1788513</a>
<p id="display">
  <iframe id="f"></iframe>
</p>
<div id="content" style="display: none">
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 1788513  **/
SimpleTest.waitForExplicitFinish();

function runTest() {
  let didThrowError = false;
  try {
    location.href = "http:// example.com";
  } catch (e) {
    ok(e instanceof TypeError, "Malformed URLs should throw TypeError");
    didThrowError = true;
  }
  ok(didThrowError, "Malformed URLs should cause an error");
  SimpleTest.finish();
}

addLoadEvent(runTest);
</script>
</pre>
</body>
</html>
